Added clip criteria expression support. Evaluate to true to clip the encounter.
Relying on roslyn to evaluate the expression. If it fails to compile you will know when the program starts.

It only supports the data in ClipCriteriaContext.

The env will default to 1 second minimum duration and no padding, so if you omit this information, you can still provide a clip criteria expression.

The old behavior still works, if you don't provide a clip expression you can keep providing a minimum duration.

Optimized parsing to not rely on a string replacement when not needed, instead, if the TryParseDefault fails, then attempt the more expensive variant, as it is rare and less frequent.

Added a general UnhandledException listener to catch any uncaught errors and print them to the console.