Added better unhandled exception handling.
In addition to printing a message, if terminating, it attempts to write it to a log file in the executable folder.

If that also fails, it simply pauses the console and awaits a key press to exit, so we can inspect the error before the window disappears.

Upgraded nuget package.